+++ tests/IntegrationTests/DI/Fixtures/wildcards.php
@@ -1,0 +1,7 @@
+<?php
+
+return array(
+    'foo*' => 'bar',
+
+    'IntegrationTests\DI\*\Interface*' => DI\object('IntegrationTests\DI\*\Implementation*'),
+);